Add the controller to do the sum (optimised by using the index) and change double to bigdecimal for amount in entity

+++ src/main/java/com/insurance/insurance_api/dto/request/ContractPatchRequest.java
@@ -3,11 +3,13 @@ package com.insurance.insurance_api.dto.request;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ContractPatchRequest {
     @JsonProperty("contract_id")
     private Long contractId;
 
     @JsonProperty("new_amount")
-    private Double newAmount;
+    private BigDecimal newAmount;
 }
